requirements.txt
asgiref==3.8.1
Django==5.0.6
django-quill-editor==0.1.40
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
